renaming latent
